+++ crates/tsz-parser/src/parser/mod.rs
@@ -36,10 +36,13 @@ mod state_diagnostics;
 mod state_exports_recovery;
 mod state_expressions;
 mod state_expressions_arrow;
+mod state_expressions_binary;
+mod state_expressions_call_member;
 mod state_expressions_literals;
 mod state_expressions_literals_object;
 mod state_expressions_literals_regex;
 mod state_expressions_tail;
+mod state_expressions_unary;
 mod state_import_attributes;
 mod state_recovery_helpers;
 mod state_statements;
